hypervisor: mmio: fix Wsign-compare by casting index to u32.
In the branch, index is 0 or positive, and num_mmio_regions is less than
max s32.

+++ hypervisor/mmio.c
@@ -209,7 +209,7 @@ void mmio_region_unregister(struct cell *cell, unsigned long start)
 		cell->mmio_generation++;
 		memory_barrier();
 
-		for (/* empty */; index < cell->num_mmio_regions; index++)
+		for (/* empty */; (u32)index < cell->num_mmio_regions; index++)
 			copy_region(cell, index + 1, index);
 
 		cell->num_mmio_regions--;
